+++ mobile/android/android-components/components/browser/state/src/test/java/mozilla/components/browser/state/action/ContentActionTest.kt
@@ -12,10 +12,12 @@ import mozilla.components.browser.state.state.SecurityInfoState
 import mozilla.components.browser.state.state.TabSessionState
 import mozilla.components.browser.state.state.content.DownloadState
 import mozilla.components.browser.state.state.content.FindResultState
+import mozilla.components.browser.state.state.content.HistoryState
 import mozilla.components.browser.state.state.createCustomTab
 import mozilla.components.browser.state.state.createTab
 import mozilla.components.browser.state.store.BrowserStore
 import mozilla.components.concept.engine.HitResult
+import mozilla.components.concept.engine.history.HistoryItem
 import mozilla.components.concept.engine.manifest.WebAppManifest
 import mozilla.components.concept.engine.prompt.PromptRequest
 import mozilla.components.concept.engine.window.WindowRequest
@@ -599,5 +601,26 @@ class ContentActionTest {
         ).joinBlocking()
 
         assertNull(tab.content.webAppManifest)
+    }
+
+    @Test
+    fun `UpdateHistoryStateAction updates history state`() {
+        val historyState = HistoryState(
+            items = listOf(
+                HistoryItem("Mozilla", "https://mozilla.org"),
+                HistoryItem("Firefox", "https://firefox.com")
+            ),
+            currentIndex = 1
+        )
+
+        assertNotEquals(historyState, tab.content.history)
+        assertNotEquals(historyState, otherTab.content.history)
+
+        store.dispatch(
+            ContentAction.UpdateHistoryStateAction(tab.id, historyState.items, historyState.currentIndex)
+        ).joinBlocking()
+
+        assertEquals(historyState, tab.content.history)
+        assertNotEquals(historyState, otherTab.content.history)
     }
 }